+++ spring-cloud-tencent-commons/src/test/java/com/tencent/cloud/common/metadata/endpoint/PolarisMetadataEndpointTests.java
@@ -16,17 +16,16 @@
  *
  */
 
-
 package com.tencent.cloud.common.metadata.endpoint;
 
 import java.util.HashMap;
 import java.util.Map;
 
 import com.tencent.cloud.common.metadata.StaticMetadataManager;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
-import org.mockito.junit.MockitoJUnitRunner;
+import org.mockito.junit.jupiter.MockitoExtension;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.when;
@@ -36,7 +35,7 @@ import static org.mockito.Mockito.when;
  *
  * @author shuiqingliu
  */
-@RunWith(MockitoJUnitRunner.class)
+@ExtendWith(MockitoExtension.class)
 public class PolarisMetadataEndpointTests {
 
 	@Mock