+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - [feature:add PolarisRateLimiterLimitedFallback spi. ](https://github.com/Tencent/spring-cloud-tencent/pull/858)
 - [fix:fix the error capture of rate limit exception.](https://github.com/Tencent/spring-cloud-tencent/pull/861)
 - [feat:enable stat reporting as default.](https://github.com/Tencent/spring-cloud-tencent/pull/864)
+- [refactor:update to junit 5.](https://github.com/Tencent/spring-cloud-tencent/pull/867)